MDEV-7486: Condition pushdown from HAVING into WHERE
Condition can be pushed from the HAVING clause into the WHERE clause
if it depends only on the fields that are used in the GROUP BY 'grouping_fields'
or depends on the fields that are equal to 'grouping_fields'.
Aggregate functions can't be pushed down.

How the pushdown is made on the example:

SELECT t1.a,MAX(t1.b)
FROM t1
GROUP BY t1.a
HAVING (t1.a>2) AND (MAX(c)>12);

=>

SELECT t1.a,MAX(t1.b)
FROM t1
WHERE (t1.a>2)
GROUP BY t1.a
HAVING (MAX(c)>12);

The implementation scheme:

1. Search for the condition cond in the HAVING clause of the select that
   depends only on the fields that are used in the GROUP BY of the select
2. Save it as a condition that can be pushed into the WHERE clause
   of the select
3. Remove cond from the HAVING clause

The optimization is made in the
st_select_lex::pushdown_from_having_into_where() and is controlled by the
variable 'condition_pushdown_from_having_into_where'.

New test file having_cond_pushdown.test is created.

optimize_cond() for the HAVING clause was moved so now it is called before
optimize_cond() call for the WHERE clause in JOIN::optimize_inner().
check_simple_equality() was also changed in the way that after the work of
optimize_cond() for the HAVING clause all equalities there are transformed into
the multiple equalities.